Refactor localization methods to new dedicated class
