+++ Source/AcknowListViewController.swift
@@ -97,7 +97,7 @@ public class AcknowListViewController: UITableViewController {
     }
 
     func commonInit(acknowledgementsPlistPath acknowledgementsPlistPath: String?) {
-        self.title = AcknowListViewController.localizedTitle()
+        self.title = AcknowLocalization.localizedTitle()
 
         if let acknowledgementsPlistPath = acknowledgementsPlistPath {
             let parser = AcknowParser(plistPath: acknowledgementsPlistPath)
@@ -116,7 +116,7 @@ public class AcknowListViewController: UITableViewController {
 
             if (headerFooter.footer == DefaultFooterText ||
                 headerFooter.footer == DefaultFooterTextLegacy) {
-                self.footerText = AcknowListViewController.localizedCocoaPodsFooterText()
+                self.footerText = AcknowLocalization.localizedCocoaPodsFooterText()
             }
             else if (headerFooter.footer != "") {
                 self.footerText = headerFooter.footer
@@ -147,75 +147,6 @@ public class AcknowListViewController: UITableViewController {
     class func defaultAcknowledgementsPlistPath() -> String? {
         let DefaultAcknowledgementsPlistName = "Pods-acknowledgements"
         return self.acknowledgementsPlistPath(name: DefaultAcknowledgementsPlistName)
-    }
-
-
-    // MARK: - Localization
-
-    /**
-     The localized version of “Acknowledgements”.
-     You can use this value for the button presenting the `VTAcknowledgementsViewController`, for instance.
-
-     - return: The localized title.
-     */
-    public class func localizedTitle() -> String {
-        return self.localizedString(forKey: "VTAckAcknowledgements", defaultString: "Acknowledgements")
-    }
-
-    class func preferredLanguageCode() -> String? {
-        return NSLocale.preferredLanguages().first
-    }
-
-    class func localizedString(forKey key: String, defaultString: String) -> String {
-        var bundlePath = NSBundle(forClass: AcknowListViewController.self).pathForResource("AcknowList", ofType: "bundle")
-        let languageBundle: NSBundle
-
-        if let currentBundlePath = bundlePath {
-            let bundle = NSBundle(path: currentBundlePath)
-            var language = "en"
-
-            if let firstLanguage = self.preferredLanguageCode() {
-                language = firstLanguage
-            }
-
-            if let bundle = bundle {
-                let localizations = bundle.localizations
-                if localizations.contains(language) == false {
-                    language = language.componentsSeparatedByString("-").first!
-                }
-
-                if localizations.contains(language) {
-                    bundlePath = bundle.pathForResource(language, ofType: "lproj")
-                }
-            }
-        }
-
-        if let bundlePath = bundlePath {
-            let bundleWithPath = NSBundle(path: bundlePath)
-            if let bundleWithPath = bundleWithPath {
-                languageBundle = bundleWithPath
-            }
-            else {
-                languageBundle = NSBundle.mainBundle()
-            }
-        }
-        else {
-            languageBundle = NSBundle.mainBundle()
-        }
-
-        let localizedDefaultString = languageBundle.localizedStringForKey(key, value:defaultString, table:nil)
-        return NSBundle.mainBundle().localizedStringForKey(key, value:localizedDefaultString, table:nil)
-    }
-
-    class func CocoaPodsURLString() -> String {
-        return "https://cocoapods.org"
-    }
-
-    class func localizedCocoaPodsFooterText() -> String {
-        return
-            self.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: "Generated by CocoaPods")
-                + "\n"
-                + self.CocoaPodsURLString()
     }
 
 
@@ -281,7 +212,7 @@ public class AcknowListViewController: UITableViewController {
     - parameter sender: The event sender.
     */
     @IBAction public func openCocoaPodsWebsite(sender: AnyObject) {
-        let url = NSURL(string: AcknowListViewController.CocoaPodsURLString())
+        let url = NSURL(string: AcknowLocalization.CocoaPodsURLString())
         if let url = url {
             UIApplication.sharedApplication().openURL(url)
         }
@@ -356,7 +287,7 @@ public class AcknowListViewController: UITableViewController {
             label.autoresizingMask = [.FlexibleLeftMargin, .FlexibleRightMargin]
             label.userInteractionEnabled = true;
 
-            let CocoaPodsURL = NSURL(string: AcknowListViewController.CocoaPodsURLString())
+            let CocoaPodsURL = NSURL(string: AcknowLocalization.CocoaPodsURLString())
             if let CocoaPodsURL = CocoaPodsURL,
                 let CocoaPodsURLHost = CocoaPodsURL.host {
                     if footerText.rangeOfString(CocoaPodsURLHost) != nil {
