+++ Tests/AcknowLocalizationTests.swift
@@ -15,73 +15,73 @@ class AcknowLocalizationTests: XCTestCase {
 
     func testLocalizedStringReturnsDefault() {
 
-        class MockAcknowListViewController: AcknowListViewController {
+        class MockAcknowLocalization: AcknowLocalization {
             override class func preferredLanguageCode() -> String {
                 return "qq-QQ"
             }
         }
 
-        XCTAssertEqual("Acknowledgements", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
-        XCTAssertEqual("Generated by CocoaPods", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
+        XCTAssertEqual("Acknowledgements", MockAcknowLocalization.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
+        XCTAssertEqual("Generated by CocoaPods", MockAcknowLocalization.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
     }
 
     func testLocalizedStringReturnsEnglish() {
 
-        class MockAcknowListViewController: AcknowListViewController {
+        class MockAcknowLocalization: AcknowLocalization {
             override class func preferredLanguageCode() -> String {
                 return "en-GB"
             }
         }
 
-        XCTAssertEqual("Acknowledgements", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
-        XCTAssertEqual("Generated by CocoaPods", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
+        XCTAssertEqual("Acknowledgements", MockAcknowLocalization.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
+        XCTAssertEqual("Generated by CocoaPods", MockAcknowLocalization.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
     }
 
     func testLocalizedStringReturnsEnglishUS() {
 
-        class MockAcknowListViewController: AcknowListViewController {
+        class MockAcknowLocalization: AcknowLocalization {
             override class func preferredLanguageCode() -> String {
                 return "en-US"
             }
         }
 
-        XCTAssertEqual("Acknowledgements", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
-        XCTAssertEqual("Generated by CocoaPods", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
+        XCTAssertEqual("Acknowledgements", MockAcknowLocalization.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
+        XCTAssertEqual("Generated by CocoaPods", MockAcknowLocalization.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
     }
 
     func testLocalizedStringReturnsFrenchForShortLanguageCode() {
 
-        class MockAcknowListViewController: AcknowListViewController {
+        class MockAcknowLocalization: AcknowLocalization {
             override class func preferredLanguageCode() -> String {
                 return "fr"
             }
         }
 
-        XCTAssertEqual("Remerciements", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
-        XCTAssertEqual("Généré par CocoaPods", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
+        XCTAssertEqual("Remerciements", MockAcknowLocalization.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
+        XCTAssertEqual("Généré par CocoaPods", MockAcknowLocalization.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
     }
     
     func testLocalizedStringReturnsFrench() {
 
-        class MockAcknowListViewController: AcknowListViewController {
+        class MockAcknowLocalization: AcknowLocalization {
             override class func preferredLanguageCode() -> String {
                 return "fr-FR"
             }
         }
 
-        XCTAssertEqual("Remerciements", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
-        XCTAssertEqual("Généré par CocoaPods", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
+        XCTAssertEqual("Remerciements", MockAcknowLocalization.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
+        XCTAssertEqual("Généré par CocoaPods", MockAcknowLocalization.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
     }
 
     func testLocalizedStringReturnsSimplifiedChinese() {
 
-        class MockAcknowListViewController: AcknowListViewController {
+        class MockAcknowLocalization: AcknowLocalization {
             override class func preferredLanguageCode() -> String {
                 return "zh-Hans"
             }
         }
 
-        XCTAssertEqual("致谢", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
-        XCTAssertEqual("使用 CocoaPods 生成", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
+        XCTAssertEqual("致谢", MockAcknowLocalization.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
+        XCTAssertEqual("使用 CocoaPods 生成", MockAcknowLocalization.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
     }
 }
